Template errors, hooray
So I've got a handle implementation. Basically, every component will
have a pointer to its parent obj, and the obj will have a vector of
handles and a GetComponent method that searches for a handle of the
given type and returns a pointer to the object. This will allow systems
to use components to get other components from their parent obj.

However, the whole thing's fucked right now because of bad template
declarations. I've gotta fix that.