Basic framework is working
No more handles and no more entities. Instead there will be systems,
which derive from the base System class and implement Update, Remove,
and Create (Create is not part of the base class). Each system will also
have a vector of Component<T>s, representing all the components of that
type.

To create a component, call the correct system's Create method. This
will take an entityId (an index for now) and a data component and create
a component wrapper (Component<T>) in the component vector using the
entityId as the index, then set m_active to true. Removing a component
sets m_active to false.

Systems are created in the ame constructor and updated in Game::Update.