Constructors, entity removal, GetComponent
I've added a Constructor class with a static constructor example. The
goal with this class is that static constructors will be added to create
the various gameplay objects.

I've added Game::RemoveEntity, which will take an entityId and loop
through all the systems attached to that ID in m_entities, calling
Remove on each of them to destroy the entity's components and finally
calling FreeVector::free on the ID to allow it to be overwritten.

GetComponent now works as intended. It may be succeeded by a
cache-friendly alternative in the future, but for now I think the
functionality will be useful.

This is it. We now have a working ECS.